Refactor Tweaks Network * Added visual toggles * Hide other item when service disabled
