+++ src/tweaks/actions.h
@@ -15,8 +15,6 @@
 #include "./appstate.h"
 #include "./values.h"
 
-const char *net_change = "/mnt/SDCARD/.tmp_update/script/network/";
-
 void action_setAppShortcut(void *pt)
 {
     ListItem *item = (ListItem *)pt;
@@ -77,157 +75,6 @@ void action_setLowBatteryAutoSave(void *pt)
 void action_setDisableStandby(void *pt)
 {
     settings.disable_standby = ((ListItem *)pt)->value == 1;
-}
-
-void action_setservicestate(const char *service_name)
-{
-    char state[256];
-    FILE *debug_file;
-    char command[512];
-    char output[512];
-
-    sync();
-
-    sprintf(state, "%supdate_networking.sh %s toggle", net_change, service_name);
-    sprintf(command, "%s 2>&1", state);
-
-    FILE *script_output = popen(command, "r");
-    if (script_output != NULL) {
-        fgets(output, sizeof(output), script_output);
-        pclose(script_output);
-    }
-
-    debug_file = fopen("/mnt/SDCARD/.tmp_update/logs/network.log", "a");
-    if (debug_file != NULL) {
-        fprintf(debug_file, "action_setservicestate: %s\n", state);
-        fclose(debug_file);
-    }
-}
-
-void action_setserviceauth(const char *service_name)
-{
-    char authed[256];
-    FILE *debug_file;
-    char command[512];
-    char output[512];
-
-    sync();
-
-    sprintf(authed, "%supdate_networking.sh %s authed", net_change, service_name);
-    sprintf(command, "%s 2>&1", authed);
-
-    FILE *script_output = popen(command, "r");
-    if (script_output != NULL) {
-        fgets(output, sizeof(output), script_output);
-        pclose(script_output);
-    }
-
-    debug_file = fopen("/mnt/SDCARD/.tmp_update/logs/network.log", "a");
-    if (debug_file != NULL) {
-        fprintf(debug_file, "action_setserviceauth: %s\n", authed);
-        fclose(debug_file);
-    }
-}
-
-void action_sethttpstate(void *pt)
-{
-    config_flag_set(".httpState", ((ListItem *)pt)->value == 1);
-    settings.http_state = ((ListItem *)pt)->value == 1;
-    action_setservicestate("http");
-}
-
-void action_setsshstate(void *pt)
-{
-    config_flag_set(".sshState", ((ListItem *)pt)->value == 1);
-    settings.ssh_state = ((ListItem *)pt)->value == 1;
-    action_setservicestate("ssh");
-}
-
-void action_setftpstate(void *pt)
-{
-    config_flag_set(".ftpState", ((ListItem *)pt)->value == 1);
-    settings.ftp_state = ((ListItem *)pt)->value == 1;
-    action_setservicestate("ftp");
-}
-
-void action_settelnetstate(void *pt)
-{
-    config_flag_set(".telnetState", ((ListItem *)pt)->value == 1);
-    settings.telnet_state = ((ListItem *)pt)->value == 1;
-    action_setservicestate("telnet");
-}
-
-void action_sethotspotstate(void *pt)
-{
-    config_flag_set(".hotspotState", ((ListItem *)pt)->value == 1);
-    settings.hotspot_state = ((ListItem *)pt)->value == 1;
-    action_setservicestate("hotspot");
-}
-
-void action_setntpstate(void *pt)
-{
-    config_flag_set(".ntpState", ((ListItem *)pt)->value == 1);
-    settings.ntp_state = ((ListItem *)pt)->value == 1;
-    action_setservicestate("ntp");
-}
-
-void action_setntpwaitstate(void *pt)
-{
-    settings.ntp_wait = ((ListItem *)pt)->value == 1;
-}
-
-void action_settelnetauthstate(void *pt)
-{
-    config_flag_set(".authtelnetState", ((ListItem *)pt)->value == 1);
-    settings.auth_telnet_state = ((ListItem *)pt)->value == 1;
-    action_setserviceauth("telnet");
-}
-
-void action_setftpauthstate(void *pt)
-{
-    config_flag_set(".authftpState", ((ListItem *)pt)->value == 1);
-    settings.auth_ftp_state = ((ListItem *)pt)->value == 1;
-    action_setserviceauth("ftp");
-}
-
-void action_sethttpauthstate(void *pt)
-{
-    config_flag_set(".authhttpState", ((ListItem *)pt)->value == 1);
-    settings.auth_http_state = ((ListItem *)pt)->value == 1;
-    action_setserviceauth("http");
-}
-
-void action_setsshauthstate(void *pt)
-{
-    config_flag_set(".authsshState", ((ListItem *)pt)->value == 1);
-    settings.auth_ssh_state = ((ListItem *)pt)->value == 1;
-    action_setserviceauth("ssh");
-}
-
-void action_wpsconnection(void *pt)
-{
-    system("sh /mnt/SDCARD/.tmp_update/script/wpsclient.sh");
-}
-
-void action_disableall(void *pt)
-{
-    settings.http_state = ((ListItem *)pt)->value == 1;
-    settings.ssh_state = ((ListItem *)pt)->value == 1;
-    settings.ftp_state = ((ListItem *)pt)->value == 1;
-    settings.telnet_state = ((ListItem *)pt)->value == 1;
-    settings.hotspot_state = ((ListItem *)pt)->value == 1;
-    settings.auth_telnet_state = ((ListItem *)pt)->value == 1;
-    settings.auth_ftp_state = ((ListItem *)pt)->value == 1;
-    settings.auth_ssh_state = ((ListItem *)pt)->value == 1;
-    settings.auth_http_state = ((ListItem *)pt)->value == 1;
-    system("/mnt/SDCARD/.tmp_update/script/network/update_networking.sh disableall");
-}
-
-void action_settzselectstate(void *pt)
-{
-    config_setNumber("tzselect", ((ListItem *)pt)->value);
-    settings.tzselect_state = ((ListItem *)pt)->value;
-    action_setservicestate("ntp");
 }
 
 void action_setMenuButtonHaptics(void *pt)
